+++ builders/employment_builder.go
@@ -3,8 +3,8 @@ package builders
 import (
 	"fmt"
 
-	"github.com/adil-faiyaz98/go-builder-kit/v2/models"
-	"github.com/adil-faiyaz98/go-builder-kit/v2/pkg/builder"
+	"github.com/adil-faiyaz98/go-builder-kit/models"
+	"github.com/adil-faiyaz98/go-builder-kit/pkg/builder"
 )
 
 // EmploymentBuilder builds a Employment model