+++ builders/geo_location_builder.go
@@ -3,7 +3,7 @@ package builders
 import (
 	"fmt"
 
-	"github.com/adil-faiyaz98/go-builder-kit/v2/models"
+	"github.com/adil-faiyaz98/go-builder-kit/models"
 )
 
 // GeoLocationBuilder builds a GeoLocation model