+++ examples/generics/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/adil-faiyaz98/go-builder-kit/v2/pkg/builder"
+	"github.com/adil-faiyaz98/go-builder-kit/pkg/builder"
 )
 
 // Person is a simple struct for demonstration